Add error implementor, to add more data but avoid backwards incompat
Exposes a soup.Error that can be inspected to see what kind of error
you are dealing with, then be able to further inspect the Error()
method to see if there is any more information for you.

Avoids breaking backwards compatibility by not changing the type of
Error in the Root object, so inspecting the returned err should
still work in other projects.